pyproject.toml
[project]
name = "ai-peer-review"
version = "0.1.0"
description = "AI Peer Review - Multi-model validation system"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.9.0",
]
